+++ CHANGELOG.md
@@ -6,6 +6,7 @@ yyyy-mm-dd
 - Fixed alliance check during login.
 - Added button to set the start system back to the current location.
 - Added button to swap start and end systems. (thanks [@NevarrTivianne](https://github.com/NevarrTivianne))
+- Added number of jumps to route list.
 - Map: Added title (on mouse over) to systems.
 - Map: Connections are redrawn when a temporary connection has been added or removed.
 - Added CSRF protection for all requests that change data.